refactor/#88 admin 모듈 정리 (#89)
