+++ saerojinro-admin/src/main/java/goorm/saerojinro/admin/api/notification/application/NotificationAdminFacade.java
@@ -1,0 +1,70 @@
+package goorm.saerojinro.admin.api.notification.application;
+
+import goorm.saerojinro.admin.api.notification.exception.EmitterNotFoundException;
+import goorm.saerojinro.admin.api.notification.presentation.request.NotificationSendRequest;
+import goorm.saerojinro.domain.notification.application.EmitterQueryService;
+import goorm.saerojinro.domain.notification.application.NotificationCommandService;
+import goorm.saerojinro.domain.notification.domain.Notification;
+import goorm.saerojinro.domain.reservation.application.ReservationQueryService;
+import goorm.saerojinro.domain.reservation.domain.Reservation;
+import goorm.saerojinro.domain.user.application.UserQueryService;
+import goorm.saerojinro.infra.notification.response.NotificationSendResponse;
+import goorm.saerojinro.infra.notification.sse.NotificationSseSender;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
+
+import java.util.List;
+
+@Component
+@RequiredArgsConstructor
+public class NotificationAdminFacade {
+	private final EmitterQueryService emitterQueryService;
+	private final NotificationCommandService notificationCommandService;
+	private final UserQueryService userQueryService;
+	private final ReservationQueryService reservationQueryService;
+	private final NotificationSseSender sseSender;
+
+	@Transactional
+	public void sendNotificationByLectureId(Long lectureId, NotificationSendRequest request) {
+		List<Reservation> reservationList = reservationQueryService.getAllByLectureId(lectureId);
+		for (Reservation reservation : reservationList) {
+			Long receiverId = reservation.getUser().getId();
+			sendNotificationByReceiverId(receiverId, request);
+		}
+	}
+
+	@Transactional
+	public void sendNotificationByReceiverId(Long receiverId, NotificationSendRequest request) {
+		Notification notification = Notification.createNotification(
+			userQueryService.getById(receiverId),
+			request.title(),
+			request.contents()
+		);
+
+		SseEmitter emitter = emitterQueryService.findById(receiverId)
+			.orElseThrow(EmitterNotFoundException::new);
+
+		notificationCommandService.save(notification);
+		NotificationSendResponse message = NotificationSendResponse.from(notification);
+
+		sseSender.sendNotification(emitter, message);
+	}
+
+	@Transactional
+	public void sendNotificationAll(NotificationSendRequest request) {
+		Notification notification = Notification.createNotification(
+			null,
+			request.title(),
+			request.contents()
+		);
+		notificationCommandService.save(notification);
+		NotificationSendResponse message = NotificationSendResponse.from(notification);
+
+		List<SseEmitter> all = emitterQueryService.findAll();
+		for (SseEmitter emitter : all) {
+			sseSender.sendNotification(emitter, message);
+		}
+	}
+}
