+++ saerojinro-admin/src/main/java/goorm/saerojinro/admin/api/notification/exception/EmitterNotFoundException.java
@@ -1,0 +1,9 @@
+package goorm.saerojinro.admin.api.notification.exception;
+
+import goorm.saerojinro.common.exception.CustomException;
+
+public class EmitterNotFoundException extends CustomException {
+	public EmitterNotFoundException() {
+		super(NotificationApiExceptionCode.EMITTER_NOT_FOUND);
+	}
+}
