--- saerojinro-admin/src/main/java/goorm/saerojinro/admin/api/presentation/AdminControllerImpl.java
@@ -1,37 +1,0 @@
-package goorm.saerojinro.admin.api.presentation;
-
-import static org.springframework.http.HttpStatus.CREATED;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import goorm.saerojinro.admin.api.application.AdminFacade;
-import goorm.saerojinro.admin.api.presentation.request.AdminCreateRequest;
-import goorm.saerojinro.admin.api.presentation.request.RoleUpdateRequest;
-import goorm.saerojinro.admin.api.presentation.response.AdminPersistResponse;
-import lombok.RequiredArgsConstructor;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/admin")
-public class AdminControllerImpl implements AdminController {
-	private final AdminFacade adminFacade;
-
-	@Override
-	@PostMapping
-	public ResponseEntity<AdminPersistResponse> createAdmin(AdminCreateRequest request) {
-		AdminPersistResponse response = adminFacade.createAdmin(request);
-		return ResponseEntity.status(CREATED).body(response);
-	}
-
-	@Override
-	@PatchMapping("/users/{userId}")
-	public ResponseEntity<Void> updateRole(@PathVariable Long userId, RoleUpdateRequest request) {
-		adminFacade.updateRole(userId, request);
-		return ResponseEntity.noContent().build();
-	}
-}
