--- saerojinro-admin/src/main/java/goorm/saerojinro/admin/api/presentation/request/RoleUpdateRequest.java
@@ -1,11 +1,0 @@
-package goorm.saerojinro.admin.api.presentation.request;
-
-import goorm.saerojinro.common.domain.BaseRole;
-import lombok.Builder;
-
-@Builder
-public record RoleUpdateRequest(
-	BaseRole permission
-) {
-
-}
