--- saerojinro-admin/src/main/java/goorm/saerojinro/admin/api/presentation/response/AdminPersistResponse.java
@@ -1,15 +1,0 @@
-package goorm.saerojinro.admin.api.presentation.response;
-
-import goorm.saerojinro.domain.user.domain.User;
-import lombok.Builder;
-
-@Builder
-public record AdminPersistResponse(
-	Long id
-) {
-	public static AdminPersistResponse from(User user) {
-		return AdminPersistResponse.builder()
-			.id(user.getId())
-			.build();
-	}
-}
