--- saerojinro-admin/src/main/java/goorm/saerojinro/admin/lecture/presentation/LectureAdminControllerImpl.java
@@ -1,38 +1,0 @@
-package goorm.saerojinro.admin.lecture.presentation;
-
-import goorm.saerojinro.admin.lecture.application.LectureAdminFacade;
-import goorm.saerojinro.admin.lecture.presentation.request.LectureCreateRequest;
-import goorm.saerojinro.admin.lecture.presentation.request.LectureUpdateRequest;
-import goorm.saerojinro.admin.lecture.presentation.response.LectureCreateResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import static org.springframework.http.HttpStatus.*;
-
-@RestController
-@RequestMapping("/api/admin/lectures")
-@RequiredArgsConstructor
-public class LectureAdminControllerImpl implements LectureAdminController {
-	private final LectureAdminFacade lectureAdminFacade;
-
-	@PostMapping
-	public ResponseEntity<LectureCreateResponse> create(@RequestBody LectureCreateRequest request) {
-		LectureCreateResponse response = lectureAdminFacade.create(request);
-		return ResponseEntity.status(CREATED).body(response);
-	}
-
-	@PatchMapping("/{id}")
-	public ResponseEntity<Void> update(@PathVariable Long id,
-									   @RequestBody LectureUpdateRequest request) {
-		lectureAdminFacade.update(id, request);
-		return ResponseEntity.noContent().build();
-	}
-
-
-	@DeleteMapping("/{id}")
-	public ResponseEntity<Void> delete(@PathVariable Long id) {
-		lectureAdminFacade.delete(id);
-		return ResponseEntity.noContent().build();
-	}
-}
