--- saerojinro-admin/src/main/java/goorm/saerojinro/admin/lecture/presentation/request/LectureUpdateRequest.java
@@ -1,33 +1,0 @@
-package goorm.saerojinro.admin.lecture.presentation.request;
-
-import goorm.saerojinro.common.domain.Category;
-import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.validation.constraints.Positive;
-import lombok.Builder;
-
-import java.time.LocalDateTime;
-
-@Builder
-public record LectureUpdateRequest(
-	@Schema(description = "강의명", example = "공간지능 혁신을 통한 온오프라인 통합 경험의 미래", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-	String title,
-
-	@Schema(description = "강의 내용", example = "Part 1: 온오프라인 경험을 연결하는 네이버 지도가 공간지능과 만나 제공하게 될 미래 모습을 소개합니다.", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-	String contents,
-
-	@Schema(description = "강의 수용 인원", example = "100", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-	@Positive Long maxCapacity,
-
-	@Schema(description = "강의 시작 시간", example = "2025-03-01T10:00:00", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-	LocalDateTime startTime,
-
-	@Schema(description = "강의 종료 시간", example = "2025-03-01T12:00:00", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-	LocalDateTime endTime,
-
-	@Schema(description = "강의 장소", example = "온라인", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-	String location,
-
-	@Schema(description = "강의 카테고리", example = "BACKEND", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-	Category category
-) {
-}
