--- saerojinro-admin/src/main/java/goorm/saerojinro/admin/notification/exception/EmitterNotFoundException.java
@@ -1,9 +1,0 @@
-package goorm.saerojinro.admin.notification.exception;
-
-import goorm.saerojinro.common.exception.CustomException;
-
-public class EmitterNotFoundException extends CustomException {
-	public EmitterNotFoundException() {
-		super(NotificationApiExceptionCode.EMITTER_NOT_FOUND);
-	}
-}
